Portable Arduino
Use Arduino on USB flash stick and without affecting anything on your PC.

To use this feature, simply make a folder named "portable" in the application
root folder where there are "hardware", "examples", "lib" and others.
In this mode, sketch folder and file association settings are disabled in
the preferences dialog.

The sketch book folder is inside the "portable" folder. This may be confusing
especially on Mac OSX "Arduino.app".